server/requirements.txt
Django==1.6.1
Flask==0.9
Jinja2==2.6
Werkzeug==0.8.3
argparse==1.2.1
beautifulsoup4==4.1.3
distribute==0.6.24
dj-database-url==0.2.1
django-crispy-forms==1.2.3
djangorestframework==2.3.12
flake8==1.7.0
gunicorn==0.17.2
ipython==0.13.1
mmstats==0.7.2
python-dateutil==2.1
python-wordpress-xmlrpc==2.2
pytz==2012j
redis==2.7.2
requests==2.2.1
rq==0.3.5
s3cmd==1.1.0-beta3
six==1.2.0
times==0.6
wsgiref==0.1.2
